Fix store word data width
 On branch dev/indexed-ldst
 Your branch is ahead of 'origin/dev/indexed-ldst' by 1 commit.
   (use "git push" to publish your local commits)

 Changes to be committed:
	modified:   ../riscv/insns/xc_str_w.h

+++ riscv/insns/xc_str_w.h
@@ -1,7 +1,7 @@
 require_extension('x');
 
 uint64_t addr       = RS1 + RS2;
-uint16_t data       = XCRD;
+uint32_t data       = XCRD;
 
 MMU.store_uint32(addr, data);
